keras-app/requirements.txt
keras
tensorflow
flask
gevent
pillow
requests
numpy
theano